Add files via upload
It doesn't work to commit. I'd message you, but 12-6 am is when I'm up and inspired to actually do work. Also didn't do some stuff on the things since I didn't really understand any of the new stuff you taught. I'm willing to learn again if you'd like to teach me what we were suppose to learn on Wednesday. Thanks for reading and have a good day or night.